diag: Move (Diagnostics).ToTfprotov6Diagnostics() method to internal/toproto6 package
Reference: https://github.com/hashicorp/terraform-plugin-framework/issues/215
