+++ .changelog/pending.txt
@@ -1,0 +1,3 @@
+```release-note:breaking-change
+diag: Removed `Diagnostics` type `ToTfprotov6Diagnostics()` method. This was not intended for usage by provider developers.
+```
