+++ diag/diagnostics_test.go
@@ -5,7 +5,6 @@ import (
 
 	"github.com/google/go-cmp/cmp"
 	"github.com/hashicorp/terraform-plugin-framework/diag"
-	"github.com/hashicorp/terraform-plugin-go/tfprotov6"
 	"github.com/hashicorp/terraform-plugin-go/tftypes"
 )
 
@@ -515,83 +514,6 @@ func TestDiagnosticsHasError(t *testing.T) {
 
 			if got != tc.expected {
 				t.Errorf("Unexpected response: got: %t, wanted: %t", got, tc.expected)
-			}
-		})
-	}
-}
-
-func TestDiagnosticsToTfprotov6Diagnostics(t *testing.T) {
-	t.Parallel()
-
-	testCases := map[string]struct {
-		diags    diag.Diagnostics
-		expected []*tfprotov6.Diagnostic
-	}{
-		"nil": {
-			diags:    nil,
-			expected: nil,
-		},
-		"Diagnostic-SeverityInvalid": {
-			diags: diag.Diagnostics{
-				invalidSeverityDiagnostic{},
-			},
-			expected: []*tfprotov6.Diagnostic{
-				{
-					Detail:   invalidSeverityDiagnostic{}.Detail(),
-					Severity: tfprotov6.DiagnosticSeverityInvalid,
-					Summary:  invalidSeverityDiagnostic{}.Summary(),
-				},
-			},
-		},
-		"Diagnostic": {
-			diags: diag.Diagnostics{
-				diag.NewErrorDiagnostic("one summary", "one detail"),
-				diag.NewWarningDiagnostic("two summary", "two detail"),
-			},
-			expected: []*tfprotov6.Diagnostic{
-				{
-					Detail:   "one detail",
-					Severity: tfprotov6.DiagnosticSeverityError,
-					Summary:  "one summary",
-				},
-				{
-					Detail:   "two detail",
-					Severity: tfprotov6.DiagnosticSeverityWarning,
-					Summary:  "two summary",
-				},
-			},
-		},
-		"DiagnosticWithPath": {
-			diags: diag.Diagnostics{
-				diag.NewAttributeErrorDiagnostic(tftypes.NewAttributePath(), "one summary", "one detail"),
-				diag.NewAttributeWarningDiagnostic(tftypes.NewAttributePath().WithAttributeName("test"), "two summary", "two detail"),
-			},
-			expected: []*tfprotov6.Diagnostic{
-				{
-					Attribute: tftypes.NewAttributePath(),
-					Detail:    "one detail",
-					Severity:  tfprotov6.DiagnosticSeverityError,
-					Summary:   "one summary",
-				},
-				{
-					Attribute: tftypes.NewAttributePath().WithAttributeName("test"),
-					Detail:    "two detail",
-					Severity:  tfprotov6.DiagnosticSeverityWarning,
-					Summary:   "two summary",
-				},
-			},
-		},
-	}
-
-	for name, tc := range testCases {
-		name, tc := name, tc
-		t.Run(name, func(t *testing.T) {
-			t.Parallel()
-
-			got := tc.diags.ToTfprotov6Diagnostics()
-
-			if diff := cmp.Diff(got, tc.expected); diff != "" {
-				t.Errorf("Unexpected response (+wanted, -got): %s", diff)
 			}
 		})
 	}
